+++ src/app/modules/icons/index.ts
@@ -2,3 +2,4 @@ export * from './app.icons';
 export * from './atons';
 export * from './poi';
 export * from './vessels';
+export * from './waypoints';