+++ src/app/types/index.d.ts
@@ -139,7 +139,6 @@ export interface FBAppData {
   };
   racing: {
     startLine: LineString;
-    finishLine: LineString;
   };
   anchor: {
     hasApi: boolean;